requirements.txt
scikit-learn==0.22.1
datasketch==1.5.0
scipy==1.4.1
nltk==3.4.5
python-dateutil==2.8.1
pandas==1.0.1

# Development dependencies
pytest==5.4.1
flake8==3.7.9
mypy==0.761
